+++ src/bin/slick.rs
@@ -50,7 +50,12 @@ The default values are:
                 .arg(Arg::new("keymap").short('k').num_args(1))
                 .arg(Arg::new("data").short('d').num_args(1))
                 .arg(Arg::new("time").short('t').num_args(1))
-                .arg(Arg::new("elapsed").short('e').num_args(1)),
+                .arg(
+                    Arg::new("elapsed")
+                        .short('e')
+                        .num_args(1)
+                        .allow_negative_numbers(true),
+                ),
         )
         .get_matches();